+++ src/bos/server/controllers/v2/boot_set/ims.py
@@ -21,6 +21,7 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
+import logging
 
 from bos.common.clients.ims import (get_arch_from_image_data,
                                     get_ims_id_from_s3_url,
@@ -34,6 +35,10 @@ from bos.server.options import OptionsData
 from .defs import DEFAULT_ARCH
 from .exceptions import BootSetArchMismatch, BootSetError, BootSetWarning, \
                         CannotValidateBootSetArch, NonImsImage
+
+
+LOGGER = logging.getLogger(__name__)
+
 
 # Mapping from BOS boot set arch values to expected IMS image arch values
 # Omits BOS Other value, since there is no corresponding IMS image arch value
@@ -64,6 +69,7 @@ def validate_ims_boot_image(bs: JsonDict, options_data: OptionsData) -> None:
             raise BootSetError(str(err)) from err
         raise BootSetWarning(str(err)) from err
     except Exception as err:
+        LOGGER.debug(exc_type_msg(err))
         if options_data.ims_errors_fatal:
             raise BootSetError(exc_type_msg(err)) from err
         if bs_arch != 'Other':
@@ -81,6 +87,7 @@ def validate_ims_boot_image(bs: JsonDict, options_data: OptionsData) -> None:
         ims_image_arch = get_arch_from_image_data(image_data)
     except Exception as err:
         # This most likely indicates that the IMS image data we got wasn't even a dict
+        LOGGER.debug(exc_type_msg(err))
         if options_data.ims_errors_fatal:
             raise BootSetError(exc_type_msg(err)) from err
         raise BootSetWarning(str(err)) from err